windows/requirements.txt
altgraph==0.17.4
bidict==0.22.1
certifi==2023.11.17
charset-normalizer==3.3.2
h11==0.14.0
idna==3.6
natsort==8.4.0
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
pefile==2023.2.7
Pillow==10.1.0
psgtray==1.0.2
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.10
pynput==1.7.6
pyscreenrec==0.4
PyScreeze==0.1.30
PySimpleGUI==4.60.5
pystray==0.18.0
python-engineio==4.8.0
python-socketio==5.10.0
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
urllib3==2.1.0
websocket-client==1.7.0
wsproto==1.2.0
